+++ src/client/components/ac_input_email/ac_input_email.js
@@ -53,8 +53,7 @@ Template.ac_input_email.helpers({
             if( self.AC.acInstance ){
                 self.AC.acInstance.checkEmailAddress( self.$( '.ac-input-email-sub input' ).val(), { testExists: wantsNew })
                     .then(( result ) => {
-                        // only display an error message if field is not empty
-                        if( result.canonical.length && !result.ok ){
+                        if( !result.ok ){
                             self.AC.displayError( result.errors[0] );
                         }
                         self.$( '.ac-input-email-sub' ).trigger( 'ac-email-data', { ok: result.ok, email: result.canonical });